Add files via upload
The main goal of the project is to facilitate the reservation process for the interactive halls. The system owner will have the authority to perform many processes such as, adding, deleting, and updating operation on the halls. While the user of the system will be asked to inset the required information and after that will be upgraded to be able to make a reservation for a certain hall. The reservation can be for the entire hall or part of it.