Ateyq website
This project aims on designing and developing a website using HTML, CSS, PHP and JavaScript for a local store called “Ateyq”, which is a store that sells elegant products such as handmade cups and plates as well ass vintage products. The developed website will provide customers with flexibility during the processes of viewing the product and past purchase. Also, the developed website will allow customers to send their feedback or any concerns by using the designed contact page.