MyTrip System
MyTrip is a system that enables users to plan their trip in the cities of the Kingdom of Saudi Arabia to promote domestic tourism. The system offers a variety of packages based on the events that each city holds. These activities include cultural and educational events. Users will plan their trip based on the packages available.